+++ packages/core/navigation-next/src/components/presentational/LayoutManager/types.js
@@ -59,6 +59,12 @@ export type ExperimentalFeatureFlags = {|
     when nav is collapsed.
   */
   experimental_hideNavVisuallyOnCollapse: boolean,
+  /**
+    NOTE: This property is experimental and may be removed in a minor release.
+
+    Changes layout manager to accomodate horizontal global navigation across the top of the page.
+  */
+  experimental_horizontalGlobalNav: boolean,
 |};
 
 export type GetRefs = ({
@@ -90,7 +96,7 @@ export type ConnectedLayoutManagerProps = {
   ...$Exact<CollapseListeners>,
   ...$Exact<ExperimentalFeatureFlags>,
   /** The top offset value to be used in navigation */
-  topOffset?: number,
+  topOffset: number,
 
   /** Boolean value to control the shadow on GlobalNavigation */
   shouldHideGlobalNavShadow?: boolean,