+++ packages/navigation/app-navigation/src/components/AppNavigation/types.ts
@@ -1,10 +1,9 @@
-import React from 'react';
+import React, { ReactNodeArray, ReactNode } from 'react';
 
 import { AppNavigationTheme } from '../../theme';
-import { ThemedPrimaryButtonProps } from '../PrimaryButton/types';
 
 export type AppNavigationProps = {
-  primaryItems: ThemedPrimaryButtonProps[];
+  primaryItems: ReactNodeArray;
   renderAppSwitcher?: React.ComponentType<{}>;
   renderCreate?: React.ComponentType<{}>;
   renderHelp?: React.ComponentType<{}>;
@@ -13,6 +12,7 @@ export type AppNavigationProps = {
   renderProfile: React.ComponentType<{}>;
   renderSearch?: React.ComponentType<{}>;
   renderSettings?: React.ComponentType<{}>;
+  moreLabel?: ReactNode;
   theme?: AppNavigationTheme;
 };