+++ packages/navigation/app-navigation/src/components/BadgeContainer/styles.ts
@@ -1,11 +1,10 @@
-export const styles = {
-  container: {
-    position: 'relative' as const,
-    zIndex: 1,
-  },
-  badgePositioner: {
-    position: 'absolute' as const,
-    top: '-6px',
-    right: '-8px',
-  },
+export const badgeContainerCSS = {
+  pointerEvents: 'none' as const,
+  position: 'absolute' as const,
+  top: '-6px',
+  right: '-8px',
+};
+
+export const containerCSS = {
+  position: 'relative' as const,
 };