+++ packages/navigation/app-navigation/src/components/IconButton/styles.ts
@@ -1,6 +1,9 @@
 import { ThemeProps, ThemeTokens } from '@atlaskit/button/types';
 import { gridSize as gridSizeFn } from '@atlaskit/theme/constants';
+
+import { skeletonCSS } from '../../common/styles';
 import { AppNavigationTheme } from '../../theme';
+import { IconButtonSkeletonProps } from './types';
 
 const gridSize = gridSizeFn();
 
@@ -34,3 +37,17 @@ export const getIconButtonTheme = ({
     spinnerStyles,
   };
 };
+
+const buttonHeight = gridSize * 4;
+
+export const iconButtonSkeletonCSS = (
+  theme: AppNavigationTheme,
+  { marginLeft, marginRight, size }: IconButtonSkeletonProps,
+) => ({
+  borderRadius: '50%',
+  marginLeft: typeof marginLeft === 'number' ? marginLeft : `${margin.left}px`,
+  marginRight: typeof marginRight === 'number' ? marginRight : 0,
+  width: typeof size === 'number' ? size : `${buttonHeight}px}`,
+  height: typeof size === 'number' ? size : `${buttonHeight}px`,
+  ...skeletonCSS(theme),
+});