+++ packages/navigation/app-navigation/src/components/ProductHome/types.ts
@@ -1,13 +1,6 @@
-import { SerializedStyles } from '@emotion/core';
-import { ComponentType, ReactNode } from 'react';
-
-export type ProductHomeComponentProps = {
-  children: ReactNode;
-  css: SerializedStyles;
-};
+import { ComponentType } from 'react';
 
 export type ProductHomeProps = {
-  component?: ComponentType<ProductHomeComponentProps>;
   /** The product icon. Expected to be an Icon from the Atlaskit Logo package. Visible on smaller screen sizes */
   icon: ComponentType<{
     size?: 'xsmall' | 'small' | 'medium' | 'large' | 'xlarge' | undefined;
@@ -17,7 +10,6 @@ export type ProductHomeProps = {
 };
 
 export type CustomProductHomeProps = {
-  component?: ComponentType<ProductHomeComponentProps>;
   iconAlt: string;
   iconUrl: string;
   logoAlt: string;