+++ packages/navigation/app-navigation/src/components/Profile/skeleton.tsx
@@ -3,9 +3,5 @@ import React from 'react';
 import { IconButtonSkeleton } from '../IconButton/skeleton';
 
 export const ProfileSkeleton = () => (
-  <IconButtonSkeleton
-    dimension={gridSize() * 4.75}
-    marginLeft={6}
-    marginRight={6}
-  />
+  <IconButtonSkeleton marginLeft={6} marginRight={6} size={gridSize() * 4.75} />
 );