+++ packages/navigation/app-navigation/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @atlaskit/app-navigation
 
+## 0.3.2
+
+### Patch Changes
+
+- [patch][9a59c6e93b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9a59c6e93b):
+
+  Fix badge and primary items container styles
+
+## 0.3.1
+
+### Patch Changes
+
+- [patch][197aa4ed2c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/197aa4ed2c):
+
+  Use context hooks in favour of emotion-theming
+
+## 0.3.0
+
+### Minor Changes
+
+- [minor][382273ee49](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/382273ee49):
+
+  Add more behaviour
+
+## 0.2.2
+
+### Patch Changes
+
+- [patch][13f8980fb2](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/13f8980fb2):
+
+  Use emotion object style
+
 ## 0.2.1
 
 - Updated dependencies [8d0f37c23e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8d0f37c23e):